RavenDB-22513 Remove old certificates view
