+++ src/Raven.Studio/typescript/viewmodels/shell.ts
@@ -42,7 +42,6 @@ import requestExecution = require("common/notifications/requestExecution");
 import studioSettings = require("common/settings/studioSettings");
 import simpleStudioSetting = require("common/settings/simpleStudioSetting");
 import clientCertificateModel = require("models/auth/clientCertificateModel");
-import certificateModel = require("models/auth/certificateModel");
 import serverTime = require("common/helpers/database/serverTime");
 import saveGlobalStudioConfigurationCommand = require("commands/resources/saveGlobalStudioConfigurationCommand");
 import saveDatabaseStudioConfigurationCommand = require("commands/resources/saveDatabaseStudioConfigurationCommand");
@@ -62,6 +61,7 @@ import serverSettings = require("common/settings/serverSettings");
 import getLatestVersionInfoCommand = require("commands/version/getLatestVersionInfoCommand");
 import StudioSearchWithDatabaseSwitcher = require("components/shell/studioSearchWithDatabaseSelector/StudioSearchWithDatabaseSwitcher");
 import ProtractedRequestMessage = require("components/shell/partials/ProtractedRequestMessage");
+import typeUtils = require("common/typeUtils");
 
 class shell extends viewModelBase {
 
@@ -409,7 +409,7 @@ class shell extends viewModelBase {
 
         if (this.clientCertificate() && this.clientCertificate().Name) {
             const tooltipProvider = () => {
-                const dbAccessArray = certificateModel.resolveDatabasesAccess(this.clientCertificate());
+                const dbAccessArray = this.resolveDatabasesAccess(this.clientCertificate());
 
                 const allowedDatabasesText = dbAccessArray.length ?
                     dbAccessArray.map(x => `<div>
@@ -439,7 +439,7 @@ class shell extends viewModelBase {
                             <dt>Thumbprint</dt>
                             <dd><strong>${this.clientCertificate().Thumbprint}</strong></dd>
                             <dt><span>Security Clearance</span></dt>
-                            <dd><strong>${certificateModel.clearanceLabelFor(this.clientCertificate().SecurityClearance)}</strong></dd>
+                            <dd><strong>${this.certificateClearanceLabel(this.clientCertificate().SecurityClearance)}</strong></dd>
                             <dt><span>Access to databases:</span></dt>
                             <dd><span>${allowedDatabasesText}</span></dd>
                             ${twoFactorPart}
@@ -468,6 +468,41 @@ class shell extends viewModelBase {
             console.error(e);
             messagePublisher.reportWarning("Failed to load routed module!", e);
         };
+    }
+
+    private certificateClearanceLabel(clearance: Raven.Client.ServerWide.Operations.Certificates.SecurityClearance) {
+        switch (clearance) {
+            case "ClusterNode":
+                return "Cluster Node";
+            case "ClusterAdmin":
+                return "Cluster Admin";
+            case "Operator":
+                return "Operator";
+            case "ValidUser":
+                return "Valid User";
+            default:
+                return "Unknown";
+        }
+    }
+
+    // copied from certificateModel.ts
+    private  resolveDatabasesAccess(certificateDefinition: Raven.Client.ServerWide.Operations.Certificates.CertificateDefinition): databaseAccessInfo[] {
+        let dbAccessInfo;
+        
+        switch (certificateDefinition.SecurityClearance) {
+            case "ClusterAdmin":
+            case "Operator":
+            case "ClusterNode":
+                dbAccessInfo =  { All : "Admin"};
+                break;
+            default:
+                dbAccessInfo = certificateDefinition.Permissions;
+        }
+        
+        const dbAccessArray = Object.entries(dbAccessInfo ?? []).map(([dbName, accessLevel]) => 
+            ({ accessLevel: `Database${accessLevel}` as databaseAccessLevel,  dbName: dbName }));
+        
+        return typeUtils.sortBy(dbAccessArray, x => x.dbName.toLowerCase());
     }
     
     private getExpirationDurationClass() {
